bugfix: Update missing translations for case types
The code changes add a new function `stringToTranslateId` to the `WaitingRoomBaseComponent` in order to convert a string to a translate ID format. This function replaces spaces and special characters with hyphens and converts the string to lowercase. The function is used to generate translate IDs for various strings in the application.