VIH-10652 add state management with NGRX (#2145)
* WIP add actions and reducers for conference

* feat: Update conference actions to include endpoint status updates and endpoint list modifications

* add tests

* feat: Add conference status to video web service and VHConference model

* add more tests for reducer functions

* feat: Add upsertPexipParticipant action to conference reducer

* feat: Update room and participant information in conference reducer

The code changes in this commit update the conference reducer to handle room and participant information. It adds the ability to update a room and its associated participants, as well as transfer participants between rooms. The changes also include updating the available rooms list and ensuring that the room information is correctly reflected in the participants and endpoints.

* subscribe to active conference in the base waiting room

* refactor: Update participant and room information handling in conference reducer
migrate to ngrx models for participant list component and child components

* handle removal of pexip participant
retain pexip info between participant updates and conference udpates

* disable linting fules for test setup class since it does not honour rules

* add new rule to ignore tests set file