VIH-10734 Bugfix claims and reducers (#2155)
* bugfix: roles not refreshing when user starts with no roles because the cache saves the user with zero roles and fails to map empty roles.

the EventHubClient.cs is also removing the claims cache using the wrong key

* bugfix: Update conference reducer to handle room and participant information

The code changes in this commit update the conference reducer to handle room and participant information. It adds the ability to update a room and its associated participants, as well as transfer participants between rooms. The changes also include updating the available rooms list and ensuring that the room information is correctly reflected in the participants and endpoints.